Add user auth

+++ resolvers/utilResolvers.js
@@ -1,10 +1,15 @@
 const mongoose = require('mongoose')
+const { AuthenticationError } = require('apollo-server-errors')
 
 const Film = require('../models/film')
 const Episode = require('../models/episode')
 
-const dropCollections = async () => {
-  console.log('Dropping collections')
+const dropCollections = async (context) => {
+  const currentUser = context.currentUser
+
+  if (!currentUser) {
+    throw new AuthenticationError("not authenticated")
+  }
 
   await Film.deleteMany({})
   await Episode.deleteMany({})
@@ -13,7 +18,7 @@ const dropCollections = async () => {
 
 const utilResolvers = {
   Mutation: {
-    dropCollections: async() => dropCollections()
+    dropCollections: async(root, args, context) => dropCollections(context)
   }
 }
